# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Electronics Retail ERP & POS System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reports/sales-by-day:
    get:
      operationId: getSalesByDay
      tags: [reports]
      summary: Daily sales totals for charting
      parameters:
        - name: days
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Daily sales data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DailySales"
  /reports/top-products:
    get:
      operationId: getTopProducts
      tags: [reports]
      summary: Top selling products by quantity
      parameters:
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
        - name: fromDate
          in: query
          schema:
            type: ["string", "null"]
        - name: toDate
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Top products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopProduct"
  /reports/gst:
    get:
      operationId: getGstReport
      tags: [reports]
      summary: GST report (CGST, SGST, IGST breakdown)
      parameters:
        - name: fromDate
          in: query
          schema:
            type: ["string", "null"]
        - name: toDate
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: GST report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GstReport"
  /reports/profit-loss:
    get:
      operationId: getProfitLoss
      tags: [reports]
      summary: Profit & loss summary for a date range
      parameters:
        - name: fromDate
          in: query
          schema:
            type: ["string", "null"]
        - name: toDate
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Profit/loss report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfitLossReport"
  /reports/category-sales:
    get:
      operationId: getCategorySales
      tags: [reports]
      summary: Sales breakdown by category
      parameters:
        - name: fromDate
          in: query
          schema:
            type: ["string", "null"]
        - name: toDate
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Category sales breakdown
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategorySales"

  # ─── CUSTOMERS ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    DailySales:
      type: object
      properties:
        date:
          type: string
        revenue:
          type: number
        salesCount:
          type: integer
        profit:
          type: number
      required: [date, revenue, salesCount, profit]
    TopProduct:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        sku:
          type: string
        categoryName:
          type: string
        quantitySold:
          type: integer
        revenue:
          type: number
      required: [productId, productName, sku, categoryName, quantitySold, revenue]
    GstReport:
      type: object
      properties:
        totalSales:
          type: number
        totalCgst:
          type: number
        totalSgst:
          type: number
        totalIgst:
          type: number
        totalTax:
          type: number
        transactionCount:
          type: integer
      required: [totalSales, totalCgst, totalSgst, totalIgst, totalTax, transactionCount]
    ProfitLossReport:
      type: object
      properties:
        revenue:
          type: number
        costOfGoods:
          type: number
        grossProfit:
          type: number
        totalExpenses:
          type: number
        netProfit:
          type: number
        grossMargin:
          type: number
        netMargin:
          type: number
      required: [revenue, costOfGoods, grossProfit, totalExpenses, netProfit, grossMargin, netMargin]
    CategorySales:
      type: object
      properties:
        categoryId:
          type: integer
        categoryName:
          type: string
        quantitySold:
          type: integer
        revenue:
          type: number
        percentage:
          type: number
      required: [categoryId, categoryName, quantitySold, revenue, percentage]

    # Customers
...